requirements.txt
Flask==0.11.1
Flask-SQLAlchemy==2.1
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.1.3
Werkzeug==0.11.11
beautifulsoup4==4.5.1
click==6.6
django-htmlmin==0.10.0
gunicorn==19.6.0
html5lib==0.999999999
itsdangerous==0.24
lxml==3.6.4
numpy==1.11.2
pandas==0.19.1
pandas-datareader==0.2.1
python-dateutil==2.5.3
pytz==2016.7
requests==2.11.1
requests-file==1.4.1
six==1.10.0
webencodings==0.5
yapf==0.14.0
